refactor: LogtoRequest
