+++ Sources/Logto/Protocols/NetworkSession.swift
@@ -10,11 +10,6 @@ import Foundation
 // https://www.swiftbysundell.com/articles/mocking-in-swift/#complete-mocking
 protocol NetworkSession {
     func loadData(
-        with url: URL,
-        completion: @escaping HttpCompletion<Data>
-    )
-
-    func loadData(
         with request: URLRequest,
         completion: @escaping HttpCompletion<Data>
     )
