+++ Sources/Logto/Utilities/LogtoUtilities+IdToken.swift
@@ -15,7 +15,7 @@ extension LogtoUtilities {
     /// - Parameter token: The JWT to decode
     /// - Returns: A set of ID Token claims
     static func decodeIdToken(_ idToken: String) throws -> IdTokenClaims {
-        let decoder = Utilities.getCamelCaseDecoder()
+        let decoder = LogtoUtilities.getCamelCaseDecoder()
 
         let segments = idToken.split(separator: ".")
